Update apk
Forgot that I had to update the apk on my side before directly pushing to github.

The program should be updated with the previous features in the last 2 commits.

Copy feature has been removed as it clashes with some parts of the pygbag module, meaning that I cant run it and may find an alternative